Add GeoStyler and move TileMatrix interface
