+++ GIFrameworkMaps.Web/package.json
@@ -33,6 +33,7 @@
     "dompurify": "3.1.5",
     "fontoxpath": "3.32.0",
     "fuse.js": "7.0.0",
+    "geostyler-openlayers-parser": "5.0.0",
     "jspdf": "2.5.1",
     "lodash": "4.17.21",
     "luxon": "3.4.4",
